Gawron/zpi 42 be 40 reservation structure

+++ src/main/java/com/example/petbuddybackend/dto/offer/OfferConfigurationDTO.java
@@ -2,6 +2,7 @@ package com.example.petbuddybackend.dto.offer;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -9,7 +10,7 @@ import java.util.Map;
 public record OfferConfigurationDTO(
         Long id,
         String description,
-        Double dailyPrice,
+        BigDecimal dailyPrice,
         Map<String, List<String>> selectedOptions //it would be easier to get list of selected options id for
                                                   // persisting/editing, but map is better structure for presentation
 ) {
